refactor(cli): extract graph group + shared cli_common (3a slice 1)
First concrete slice of the cli.py (9.3k-line) decomposition, proving the
pattern the rest of the groups will follow.

- cli_common.py: shared CLI state (the rich Console + get_memory factory)
  lifted out so group modules don't import the 9k-line cli.py back — that
  would be a circular import. cli.py now imports console/_get_memory from
  here (zero call-site churn — kept the _get_memory alias).
- cli_graph.py: the `memo graph` group (path/neighbors/communities/
  centrality/export) as a standalone @click.group, registered onto the
  root via cli.add_command(graph_group).

cli.py -175 lines. `memo graph --help` works; suite 732 green, mypy + ruff
clean. Remaining groups (~29) follow this same extract-and-register shape.
